+++ R/sarb_code.R
@@ -39,5 +39,5 @@ sarb_code <- function(code, token = NULL){
   res <- jsonlite::fromJSON(content(res, "text", encoding = "UTF-8"))
 
   res %>%
-    tbl_df
+    as_tibble()
 }