+++ NEWS.md
@@ -1,3 +1,8 @@
+# sarbR 0.1.1
+
+* Addition of function `release_info` to get information on latest release date of quaterly data in database
+* Fix bug in `sarb_code` not picking up token
+
 # sarbR 0.1.0
 
 * The 1 second sleep between requests have been removed